+++ src/main/java/de/alpharogroup/swing/laf/actions/LookAndFeelAction.java
@@ -35,8 +35,7 @@ import de.alpharogroup.swing.laf.LookAndFeels;
 /**
  * The class {@link LookAndFeelAction}.
  * 
- * @deprecated use instead the same name class in the new package de.alpharogroup.swing.plaf
- *             <br>
+ * @deprecated use instead the same name class in the new package de.alpharogroup.swing.plaf <br>
  *             Note: will be removed in the next minor release
  */
 public class LookAndFeelAction extends AbstractAction