+++ src/main/java/de/alpharogroup/swing/laf/actions/LookAndFeelMetalAction.java
@@ -39,8 +39,7 @@ import de.alpharogroup.swing.laf.LookAndFeels;
 /**
  * The class {@link LookAndFeelMetalAction}.
  * 
- * @deprecated use instead the same name class in the new package de.alpharogroup.swing.plaf
- *             <br>
+ * @deprecated use instead the same name class in the new package de.alpharogroup.swing.plaf <br>
  *             Note: will be removed in the next minor release
  */
 public class LookAndFeelMetalAction extends LookAndFeelAction